Modify test and source code to pass tests
Makefile is modified to compile SOURCES instead of TARGET against
tests target. Makefile design needs to be investigated further.
minunit.h is modified to accept variable arguments for printing
values in error messages.
All tests pass except for final test which tests if the stack == NULL.
This test has been commented out for now. The program has been run
seperately in Valgrind, so the memory is not an issue. The test is
required simply as a confirmation that the pointer is set to NULL.
The reason for the test failing when the pointer is set to NULL in
Stack_destroy() must be investigated.